clean up some method names
